+++ spec/qac_spec.rb
@@ -97,7 +97,7 @@ describe "Qac" do
   it 'integration test' do
     if $qacInstalled
 
-      exit_code = Bake.startBakeqac("qac/main", ["test_template", "--qacretry", "60", "--qacdoc"])
+      exit_code = Bake.startBakeqac("qac/main/src", ["test_template", "--qacretry", "60", "--qacdoc"])
 
       $mystring.gsub!(/\\/,"/")